chore: change call api and add Readme

+++ server/src/controllers/UserController.ts
@@ -8,16 +8,16 @@ class UserController {
     const repository = getRepository(User)
     const { id, nome, idade, estadoCivil, cpf, cidade, estado } = req.body
 
-    const userExists = await repository.findOne(cpf)
+    const userExists = await repository.find({ cpf: cpf })
 
-    if (userExists) {
-      return res.status(409).json({ success: false, message: 'O cpf já foi cadastrado' })
+    if (userExists.length === 1) {
+      return res.status(201).json({ success: false, message: 'O cpf já foi cadastrado' })
     } else {
       const user = repository.create({ id, nome, idade, estadoCivil, cpf, cidade, estado })
 
       await repository.save(user)
 
-      return res.status(201).json({ success: true, message: 'Usuario criado com sucesso!', data: user })
+      return res.status(201).json({ success: true, message: 'Usuário criado com sucesso!', data: user })
     }
   }
 
@@ -26,17 +26,19 @@ class UserController {
     const limit = Number(req.query.limit)
 
     const repository = getRepository(User)
-    const users = await repository.find()
+    const [users, totalUsers] = await repository.findAndCount()
 
     const startIndex = (page - 1) * limit
     const endIndex = page * limit
 
+    const totalPage = totalUsers / limit
+
     if (users) {
       const userResult = users.slice(startIndex, endIndex)
 
-      return res.status(201).json({ success: true, message: 'Usuarios encontrados!', data: userResult })
+      return res.status(201).json({ success: true, message: 'Usuários encontrados!', data: userResult, totalPages: totalPage })
     } else {
-      return res.status(201).json({ success: true, message: 'Nenhum usuario cadastrado' })
+      return res.status(201).json({ success: true, message: 'Nenhum usuário cadastrado', data: [] })
     }
   }
 
@@ -46,7 +48,7 @@ class UserController {
 
     const user = await repository.findOne(id)
 
-    return res.status(201).json({ sucess: true, message: 'Usuario encontrado com sucesso', data: user })
+    return res.status(201).json({ success: true, message: 'Usuário encontrado com sucesso', data: user })
   }
 
   async updateUser (req:Request, res:Response) {
@@ -57,10 +59,14 @@ class UserController {
 
     let user = await repository.findOne(id)
 
-    user = { id, nome, idade, estadoCivil, cpf, cidade, estado }
-    await repository.save(user)
+    if (user) {
+      user = { id, nome, idade, estadoCivil, cpf, cidade, estado }
+      await repository.save(user)
 
-    return res.status(201).json({ sucess: true, message: 'Usuario atualizado com sucesso', user })
+      return res.status(201).json({ success: true, message: 'Usuário atualizado com sucesso', user })
+    } else {
+      return res.status(201).json({ success: false, message: 'Usuário não encontrado' })
+    }
   }
 
   async deleteUser (req:Request, res:Response) {
@@ -69,9 +75,13 @@ class UserController {
     const id = parseInt(req.params.id)
 
     const user : any = await repository.findOne(id)
-    await repository.remove(user)
+    if (user) {
+      await repository.remove(user)
 
-    return res.status(201).json({ user })
+      return res.status(201).json({ success: true, message: 'Usuário deletado com sucesso' })
+    } else {
+      return res.status(201).json({ success: false, message: 'Usuário não encontrado' })
+    }
   }
 }
